+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 # Change Log
 
 ## Unreleased
- - Updated partition key handling when creating collections to allow for leading '/' characters in the partition key - fixes [Issue #153]((https://github.com/PlagueHO/CosmosDB/issues/153)
+
+- Updated partition key handling when creating collections to allow for
+  leading '/' characters in the partition key - fixes [Issue #153](https://github.com/PlagueHO/CosmosDB/issues/153)
+- Add support for setting URI and Key when using with a Cosmos DB
+  Emulator - fixes [Issue #155](https://github.com/PlagueHO/CosmosDB/issues/155)
 
 ## 2.1.5.548